SOFTWARE REQUIREMENT SPECIFICATION
The purpose of this document is to present a detailed description of the software “IPMS” (Installment Product Management System). It will explain the purpose and features of the software, the interfaces of the software, what the software will do and the constraints under which it must operate. This document is intended for users of the software and also potential developers